requirements.txt
gradio
requests
sentence-transformers
faiss-cpu
python-docx
pydantic
html2text
